pyproject.toml
[tool.poetry]
name = "dalle-st"
version = "0.1.0"
description = ""
authors = ["p.kandapagari <pavan.kandapagari@agile-robots.com>"]
readme = "README.md"
packages = [{include = "dalle"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27"
gradio = "^3.36"
python-dotenv = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
